refactor: extract ConceptSetsTable from ConceptSets
Move MRT setup, search state, column definitions, and header JSX
into a dedicated ConceptSetsTable component. ConceptSets is now a
thin orchestrator (~90 lines) handling data fetching, tabs, and the
delete dialog only.
